refactor: divide types in graphql folder into separate folders
