+++ src/components/pages/EmployeesCVsPage/EmployeesCVsPage.tsx
@@ -4,11 +4,11 @@ import React, { useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useUser } from '../../../hooks/useUser';
 import { UserRoles } from '../../../constants/userRoles';
-import { IUserAllResult } from '../../../interfaces/IUser.interface';
 import { USER } from '../../../graphql/queries/user';
 import { RoutePath } from '../../../constants/routeVariables';
 import { Spinner } from '../../Spinner';
 import { PrivateButton } from '../../UI/PrivateButton';
+import { IUserAllResult } from '../../../graphql/types/results/userTypeResult';
 import { CvsList } from './CvsList/CvsList';
 import { Row } from './Row/Row';
 import { CvEditModal } from './CvEditModal/CvEditModal';
