+++ src/components/pages/EmployeesPage/EmployeesModal/EmployeesModal.tsx
@@ -3,11 +3,9 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import React, { FC } from 'react';
 import { useForm } from 'react-hook-form';
 import { FieldNameEmployeesForm } from '../../../../constants/FieldNameEmployeesForm';
-import { CREATE_USER } from '../../../../graphql/mutations/createUser/createUser';
-import { CreateUserResult } from '../../../../graphql/mutations/createUser/createUser.types';
-import { updateCacheAfterCreatingUser } from '../../../../graphql/mutations/createUser/createUserUpdateCache';
+import { CREATE_USER } from '../../../../graphql/mutations/createUser';
+import { updateCacheAfterCreatingUser } from '../../../../graphql/cache/createUser.cache';
 import { useEmployeesFormData } from '../../../../hooks/useEmployeesFormData';
-import { IUserAllResult } from '../../../../interfaces/IUser.interface';
 import { TError } from '../../../../types/errorTypes';
 import { TFormSubmit } from '../../../../types/formTypes';
 import { employeesSchema } from '../../../../utils/validationSchema';
@@ -15,6 +13,7 @@ import { Spinner } from '../../../Spinner';
 import { InputSelect } from '../../../UI/InputSelect';
 import { InputText } from '../../../UI/InputText';
 import { ModalWindow } from '../../../UI/ModalWindow';
+import { CreateUserResult, IUserAllResult } from '../../../../graphql/types/results/userTypeResult';
 import { IEmployeesFormInput, IEmployeesModalProps } from './EmployeesModal.interface';
 import * as Styled from './EmployeesModal.styles';
