+++ src/components/pages/EmployeesProfilePage/AvatarModal/AvatarModal.tsx
@@ -3,16 +3,16 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { useForm } from 'react-hook-form';
 import { useMutation, useQuery } from '@apollo/client';
 import CancelIcon from '@mui/icons-material/Cancel';
-import { IUserAllResult } from '../../../../interfaces/IUser.interface';
 import { USER } from '../../../../graphql/queries/user';
-import { IAvatarReturn } from '../../../../interfaces/IAvatar';
-import { DELETE_AVATAR, UPLOAD_AVATAR } from '../../../../graphql/mutations/avatar/avatar';
-import { updateUserCacheAfterAvatarMutation } from '../../../../graphql/mutations/avatar/avatar.cache';
+import { IAvatarReturn } from '../../../../graphql/types/results/avatarTypeResult';
+import { DELETE_AVATAR, UPLOAD_AVATAR } from '../../../../graphql/mutations/avatar';
+import { updateUserCacheAfterAvatarMutation } from '../../../../graphql/cache/avatar.cache';
 import { avatarSchema } from '../../../../utils/validationSchema';
 import { TError } from '../../../../types/errorTypes';
 import { convertToBase64 } from '../helpers/convertToBase64';
 import { ModalWindow } from '../../../UI/ModalWindow';
 import { Spinner } from '../../../Spinner';
+import { IUserAllResult } from '../../../../graphql/types/results/userTypeResult';
 import * as Styled from './AvatarModal.styles';
 import { InputFile } from './InputFile/InputFile';
 import { IAvatarForm, IAvatarModal } from './AvatarModal.types';
