+++ src/components/pages/EmployeesProfilePage/EmployeesProfilePage.tsx
@@ -4,8 +4,8 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { RoutePath } from '../../../constants/routeVariables';
 import { UserRoles } from '../../../constants/userRoles';
 import { USER } from '../../../graphql/queries/user';
+import { IUserAllResult } from '../../../graphql/types/results/userTypeResult';
 import { useUser } from '../../../hooks/useUser';
-import { IUserAllResult } from '../../../interfaces/IUser.interface';
 import { chooseAvatarLetter } from '../../../utils/chooseAvatarLetter';
 import { Spinner } from '../../Spinner';
 import { PrivateButton } from '../../UI/PrivateButton';
