+++ src/graphql/cache/createUser.cache.ts
@@ -1,7 +1,7 @@
 import { ApolloCache, NormalizedCacheObject } from '@apollo/client';
-import { IUsersResult } from '../../../components/pages/EmployeesPage/EmployeesPage.interface';
-import { GET_ALL_USERS } from '../../queries/users';
-import { CreateUserResult } from './createUser.types';
+import { IUsersResult } from '../../components/pages/EmployeesPage/EmployeesPage.interface';
+import { GET_ALL_USERS } from '../queries/users';
+import { CreateUserResult } from '../types/results/userTypeResult';
 
 export const updateCacheAfterCreatingUser = (
   cache: ApolloCache<NormalizedCacheObject>,
