+++ src/graphql/cache/deleteUser.cache.ts
@@ -1,5 +1,5 @@
 import { ApolloCache } from '@apollo/client';
-import { IUser } from '../../../interfaces/IUser.interface';
+import { IUser } from '../../interfaces/IUser.interface';
 
 export const updateCacheAfterDeleteUser = (cache: ApolloCache<IUser>, userId: string) => {
   const id = cache.identify({ id: userId, __typename: 'User' });
