+++ src/graphql/mutations/avatar.ts
