--- src/graphql/mutations/createUser/createUser.types.ts
@@ -1,17 +1,0 @@
-import { IProfileInput } from '../../../interfaces/IProfile.interface';
-import { IUser } from '../../../interfaces/IUser.interface';
-import { IAuthInput } from '../../authentication/authResult.interface.';
-
-export type CreateUserInput = {
-  user: {
-    auth: IAuthInput;
-    profile: IProfileInput;
-    cvsIds: string[];
-    departmentId: string;
-    positionId: string;
-  };
-};
-
-export type CreateUserResult = {
-  createUser: IUser;
-};
