+++ src/graphql/mutations/deleteUser.ts
