--- src/graphql/mutations/deleteUser/deleteUser.types.ts
@@ -1,13 +1,0 @@
-export type DeleteUserInput = {
-  id: string;
-};
-
-export type DeleteUserResult = {
-  deleteUser: {
-    affected: number;
-  };
-};
-
-export type DeleteResult = {
-  affected: number;
-};
