+++ src/graphql/types/inputs/profileTypeInput.ts
@@ -1,0 +1,9 @@
+import { ILanguageProficiency } from '../../../interfaces/ILanguageProficiency.interface';
+import { ISkillMastery } from '../../../interfaces/ISkillMastery.interface';
+
+export interface IProfileInput {
+  first_name: string;
+  last_name: string;
+  skills: ISkillMastery[];
+  languages: ILanguageProficiency[];
+}
