+++ src/graphql/types/results/avatarTypeResult.ts
