+++ src/graphql/types/results/cvTypeResult.ts
@@ -1,0 +1,9 @@
+import { ICv } from '../../../interfaces/ICv.interface';
+
+export interface ICvResult {
+  updateCv: ICv;
+}
+
+export interface ICvUnbindResult {
+  unbindCv: ICv;
+}
