+++ src/graphql/types/results/departmentTypeResult.ts
@@ -1,0 +1,5 @@
+import { IDepartment } from '../../../interfaces/IDepartment.interface';
+
+export interface IDepartmentReturn {
+  departments: IDepartment[];
+}
