+++ src/graphql/types/results/positionTypeResult.ts
@@ -1,0 +1,5 @@
+import { IPosition } from '../../../interfaces/IPosition.interface';
+
+export interface IPositionReturn {
+  positions: IPosition[];
+}
