+++ src/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useReactiveVar } from '@apollo/client';
-import { authService } from '../graphql/authentication/authService';
+import { authService } from '../graphql/service/authentication/authService';
 
 export const useAuth = () => {
   const token = useReactiveVar(authService.access_token$);
