+++ src/hooks/useUser.ts
@@ -1,5 +1,5 @@
 import { useReactiveVar } from '@apollo/client';
-import { authService } from '../graphql/authentication/authService';
+import { authService } from '../graphql/service/authentication/authService';
 
 export const useUser = () => {
   const user = useReactiveVar(authService.user$);
