+++ src/interfaces/ICv.interface.ts
@@ -14,11 +14,3 @@ export interface ICv {
   languages: ILanguageProficiency[];
   is_template: boolean;
 }
-
-export interface ICvResult {
-  updateCv: ICv;
-}
-
-export interface ICvUnbindResult {
-  unbindCv: ICv;
-}
