+++ src/interfaces/IUser.interface.ts
@@ -16,19 +16,3 @@ export interface IUser {
   position_name?: string;
   role: string;
 }
-
-export interface IUserAllResult {
-  user: IUser;
-}
-
-export interface IUserName {
-  email: string;
-  profile: {
-    first_name: string | null;
-    last_name: string | null;
-  };
-}
-
-export interface IUserNameResult {
-  user: IUserName;
-}
